旧版文件上传完成

+++ server/router/internal/middleware.go
@@ -90,7 +90,7 @@ func (m *middleware) OperationRecord(r *ghttp.Request) {
 
 	if token, err := api.GfJWTMiddleware.ParseToken(r); err != nil { // 优先从jwt获取用户信息
 		id, _ := strconv.Atoi(r.Request.Header.Get("x-user-id"))
-		if m.result, m.err = service.Admin.FindAdminById(&request.GetById{Id: id}); m.err != nil {
+		if m.result, m.err = service.Admin.FindAdminById(&request.GetById{Id: uint(id)}); m.err != nil {
 			g.Log().Error(g.Map{"err": m.err})
 		}
 	} else {
